pyproject.toml
[project]
name = "mecagent-technical-test"
version = "0.1.0"
description = "Just a technical test"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cadquery>=2.5.2",
    "datasets>=3.6.0",
    "ipykernel>=6.29.5",
    "scipy>=1.15.3",
    "trimesh>=4.6.11",
    "cadquery-ocp<7.7.1",
    "torch>=2.7.1",
    "transformers>=4.52.4",
    "pillow>=11.2.1",
    "accelerate>=1.8.1",
    "qwen-vl-utils>=0.0.11",
    "torchvision>=0.22.1",
    "bitsandbytes>=0.46.0",
    "peft>=0.15.2",
    "trl>=0.19.0",
    "unsloth[torch]>=2024.8",
]
